Try to find a free file descriptor for locking
Hardcoding a value might break with package managers that use pipes to
comunicate with the bash side.

We only look as far as 9 because the bash manpage says
       Redirections using file descriptors greater than 9 should be
       used with care, as they may conflict with file descriptors the
       shell uses internally.
- it's not clear whether this can be avoided by checking whether the
FD is already open, so let's be safe.

Unfortunately, it seems bash doesn't have a way to vary the redirected
FD number except for eval. :-(