Add tokyocabinet option.
This patch adds gdbm and tokyocabinet options with an annotation to
select exactly one of the option flags.
A bit information about Tokyo Cabinet:
Tokyo Cabinet is developed as the successor of GDBM and QDBM and it's
more efficient than both gdbm and qdbm.
An exheres for tokyocabinet is available in ::alip-misc.