Fix GCC 4.4.0 -> 4.4.1 upgrade path
Ensure that libstdc++.so symlinks always point to the highest available so version.

When upgrading to a new so version in the highest slot the new so
version needs to be installed during merge. Then the symlinks need to be
updated before the old version is unmerged. pkg_postinst is too late for
this. pkg_prerm does not know which version was just merged and can't be
fixed retroactively.
Therefore we fix the links in src_install in the image (which creates
collisions). This part can be removed when everybody have upgraded their
latest gcc slot. Further we start looking for the latest so version in
pkg_prerm and rewrite to keep it updated and make sure it's never
unmerged.

Remove previous hacks added in 8e47f2f7, 84a823fa.

Signed-off-by: Ingmar Vanhassel <ingmar@exherbo.org>