dovecot: add init script and modify exheres accordingly
This init script is lifted from Gentoo.
The only modification was to remove 'net'
from the items dovecot reportedly needs.

Dovecot will run perfectly with no connection to the
outside world and in Exherbo, init scripts that are
"scheduled to start" after a wireless device establishes
its connection often never do. (sshd, openntpd)

While a (dovecot) server is unlikely to have a wireless
card as its only connection, mine currently does, and
does run into this problem, so this seemed a lesser-of-
evils approach. This is noted in comment in the init
script.

Hopefully, this whole long-winded treatise is a moot point
to begin with, and Exherbo will soon have an init system
as gloriously right as the rest of it is :)