python.exlib: Nuke python_mod_cleanup()
Now that .py(|c|o) files are part of contents, they will be unmerged
properly when the package is uninstalled.

They may be regenerated under these conditions:
- mtime of the .py changes
- the .py is run by a user with write access to the directory containing
  the .py

In that case, the .py(|c|o) won't be unmerged, but this isn't something
the exlib should worry about, and isn't any different from non-python
packages.

Signed-off-by: Ingmar Vanhassel <ingmar@exherbo.org>
Reviewed-by: Bo Ørsted Andresen <zlin@exherbo.org>