dracut: fix install paths and adjust deps
Minor dep adjustments and fixed install paths so the configuration dir
is not installed to /usr/etc.

The tests on this package still need to be fixed. The first series
simply uses dash to check the syntax of the modules, but the second
series creates an initrd and rootfs and launches an instance of qemu
to test it. This is done nine times and requires qemu, lvm, mdadm  and
various other network fs packages. This second set of tests might be a
good candidate for test-expensive.