glib: fix automagic dep on DBus
DBus is only required for a single GDBus test and it was being built
depending on the presence of DBus.

In order for eautoconf to work correctly gtk-doc must be installed since
gtk-doc m4 macros are used, so it is a dependency while this patch is
being used. As an alternative this small m4 file could be provided by us
with the patch and included when running aclocal.