glib: remove stupid dbus test patch
Patching to prevent a single test from being built if someone does not
have dbus is silly. This patch was originally to prevent dbus from being
a hard dependency. DBus will just have to be a hard dependency. It is
required during runtime/tests and to build one of the tests which are
built in the all target and not the check target.

If someone wants to move this to a recommendation and test dependencies
then they need to add a --disable-tests switch to prevent them from
being built.