cryptsetup: Don't install a build-only wrapper script as cryptsetup-static.
The /sbin/cryptsetup.static we installed before this change was actually a
wrapper script (cf. http://paste.pocoo.org/show/337925/) for building (!)
and was never meant to exist in the live fs - especially since it doesn't
actually work.
It *breaks* things trying to actually use it and, thus, potentially leaves
systems dead in the water.
Furthermore, this change makes the exheres much easier. cryptsetup's tests
are passing and, most importantly, it actually works now.