dbus: adjust console auth dir and pid file
The console auth directory must end in a slash. DBus simply appends a user name
and checks if the file exists. This would fail and break dbus policies using the
at_console dbus access control ( i.e. NetworkManager clients ).

The pid file value is substituted into the system config and systemd service.
If we want to use /run/dbus.pid and not the default /run/dbus/pid we should
give it here. This also makes the sed'ing of the system.conf unnecessary.