Echo scm command to stderr, and use edo for rm -rf
Echoing to stderr is consistent with other things that echo a command
and then execute it, and also means we can show what we're doing in
appraise functions (whose stdout is captured) where appropriate, for
example the next commit.

We don't use edo for the scm command because some of them have slightly
special behaviour, but use it for the rm -rf calls to simplify a little.