Upgrade svn checkout if necessary
As of svn 1.7, the user needs to upgrade the working copy explicitly if
the new version no longer supports the old format (previous versions did
this automatically).  The appraise function isn't the most
architecturally beautiful place to do this, but it's about the only way
it would work, and since it doesn't require network access it's not too
heinous.