Initial exheres for kmod-3.
(lib)kmod is the designated replacement for module-init-tools. I'm
adding it to arbor using alternatives to allow for installing it in
parallel with module-init-tools till we remove the latter.

kmod is designed to be backwards-compatible with module-init-tools
and is currently (as of version 3 introduced here) almost completely
functional apart from a few rather exotic options. These will most
likely be introduced in version 4.