autotools.exlib: grep automake --help for '--force-missing'
and enable it if supported.
I tried to find out since when this option is supported but
didn't find any hard facts.

this is also what upstream is doing in autoreconf (2.63),
so they don't seem to know either ;-)

We don't use autotools_run_tool here, because we don't want all the
noise it generates, like selected versions, but the clean --help output.