bump parted to 3.1
libuuid moved from e2fsprogs to util-linux.
linux-headers is now required, but is in system
and shouldn't be depended on explicitly

Since parted 3.0, filesystem-related features, like resizing are gone.
A limited amount was added back in 3.1 (HFS+ and FAT only).

Still, all users of libparted seem to work fine, but need to be rebuilt.
I checked libvirt, udisks and gparted, specifically.