systemd, udev: follow upstream in using ${LIBDIR} for libexecdir
(Patch adjusted by me, WK.)

libexec is only for executables and systemd/udev recently started installing
more there. Not following this already caused problems with other upstreams
(e. g. dracut creating a broken initramfs which is potentially "not good"
for booting) and, thus, as annoying another (and hopefully final) move of
the systemd/udev rulesdir is, it needs to be done.
The underlying problem is that systemd/udev expect --libexecdir=/usr/lib
(or in our case: /use/${LIBDIR}) without complaining if it isn't.