Bring old llvm 3.0 exheres back for mesa-8.0
This ist still ugly, but not as bad as it looks.

Nothing changes if you still got this old version installed, so no revbump.

Will die with fire the moment a new mesa is out.
(which should be roughly six months after 8.0, or so they say)

If you actually want to figure out what I changed, use two checkouts,
git seems to fuck up the diff for some reason here.