nfs-utils: Fix up some init script confusion
For some reason, an out-of-date unused copy of the 'nfs' init script was
present in the files directory, delete it. (The 'nfs.initd' file is the
one that actually gets installed.)

Fix up the systemd scripts: They seem to have imported a bug from the
out of date nfs init script, and were attempting to mount the 'nfsd'
filesystem on /proc/fs/nfs. This is unnecessary; it should only be
mounted on /proc/fs/nfsd. (The /proc/fs/nfs directory is automatically
populated with the 'exports' file by the kernel.)

Revbump so people get the fixed systemd scripts.