gcc: fix dependencies for LTO
When LTO was first introduced in GCC 4.5, it required the use of
libelf[>=0.8.13], though it would also work with libelf.so from
elfutils[>=0.143].  libelf and elfutils both provide libelf.so, however, they
are incompactible with one another.

GCC 4.6+ dropped the dependency on libelf and is implicitly enabled for all
platforms which use ELF executables.

Change the dependencies accordingly for GCC 4.6+.  For GCC 4.5, force the use of
elfutils rather than libelf.