gcc: last round of refactoring; whitespace changes
This pulls out the slotting of the runtime.  It also fixes up a small issue in
the locale data slotting.  Finally, some whitespace changes and quoting changes
to round off the changes.  With this, the changes for the multiarch is about as
minimal as possible.  (That is, all the compatible changes have been ported to
master).