Revert "glibc: don't use LIBDIR"
This reverts commit 4d32c13cdf11558971a4556d7ec2bf39524ddf69.

This is a partial reversion of the aforementioned commit.  The patch was invalid
in changing all usage of ${LIBDIR}.  The usage in src_install is meant for
compatibility purposes (where ${LIBDIR} *must* be used to correctly provide the
symbolic link for the loader to avoid breaking previously compiled applications
with PT_INTERP set to /${LIBDIR}/<loader>).