glibc: enable SSP globally
This fixes an insidious bug in the static-gcc compilation.  The recursive
configure script does not pass the config.cache, resulting in the loss of
certain important configuration overrides.  Explicitly pass information through
the environment during the compile phase to ensure that the static compiler is
built appropriately for the environment.

The result of the missing parameter resulted in genemit mis-preprocessing the
generated instruction emission routines (insn-emit.c) which would fallback to
performing stack protection via the non-TLS canary setup routines.  This would
have the compiler emit a PIC call to __stack_chk_guard, resulting in undefined
relros when compiling application with -fstack-protector (e.g. libc).

With static-gcc configured correctly, we can now safely enable SSP support in
libc across all platforms.  This is a security feature with a fairly low
overhead, particularly on well supported architectures (e.g. x86, x86_64, PPC).
On such platforms, it is a TCB offset write/compare.