glibc: Only create compatibility symlinks in preinst if needed.
When /lib/ld-linux*.so* get removed cp no longer works. Calling /bin/cp with the
ldso from ${IMAGE} makes it possible to restore functionality.

We cannot symlink since the target we would symlink to is not merged yet.

This cannot be done in postinst since the merger would already have unmerged the
ldso making it impossible to execute postinst.