github.exlib: Use new hashless tag tarball downloads
Now you can just specify a tag name as the git revision
to download for a tag tarball, instead of a sha. Full
shas can still be used.

NOTE: This will break existing packages if they specify
a truncated sha to download, instead of a full sha.

Tarballs are now saved with a .tgz instead of .tar.gz
extension, purely to avoid breaking packages for users
who have an old-style tarball sitting in their distfiles
cache from a previous fetch.

Signed-off-by: Robin Green <greenrd@greenrd.org>