tcl: Fix paths in tclConfig.sh
Paths broke when multilib was added because ${WORK} no longer ended with
'/unix', so the sed command missed those variables.

Instead of a fragile sed command, patch tclConfig.sh.in, replacing
BUILD variables with the corresponding variables so ${WORK} is not
referenced.

Also, set TCL_SRC_DIR to be empty. Currently, it points to a directory
that does not exist (which I believe came from the gentoo ebuild; they
install private headers in that directory).

Dealing with DBGX in TCL_LIB_FILE is no longer necessary because:

    # TCL_DBGX used to be used to distinguish debug vs. non-debug builds.
    # This was a righteous pain so the core doesn't do that any more.

Change-Id: If37a4bb430c1aea93186e83ade77615431175894